chore(be): make search args 함수 mail-util 파일로 분리
